+++ spt-lib/src/main/java/kz/spt/lib/model/dto/payment/BillingInfoErrorDto.java
@@ -1,9 +1,13 @@
 package kz.spt.lib.model.dto.payment;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import java.math.BigDecimal;
+
 public class BillingInfoErrorDto {
 
     public String txn_id;              // уникальны номер транзакции бил-ой системы Halyk
     public int result;                 // - 1 Показатель ошибки.
-    public int sum;                    // 200 - Сумма к оплате.
+    public BigDecimal sum;                    // 200 - Сумма к оплате.
     public String message;             // "Некорректный номер авто свяжитесь с оператором."
 }
