+++ spt-lib/src/main/java/kz/spt/lib/model/dto/payment/CommandDto.java
@@ -1,9 +1,13 @@
 package kz.spt.lib.model.dto.payment;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import java.math.BigDecimal;
+
 public class CommandDto {
 
     public String command;
     public String account;
     public String txn_id;
-    public int sum;
+    public BigDecimal sum;
 }
